Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
python-dotenv==1.0.1
google-genai==0.6.0
